refactor: move solvers to separate directory
